requirements.txt
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
fonttools==4.38.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-javabridge==4.0.3
python-weka-wrapper3==0.2.11
pytz==2022.6
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
Werkzeug==2.2.2
